requirements.txt
numpy
opencv-python
matplotlib
scikit-image
scikit-learn
torch
torchvision
tensorflow
Pillow
albumentations
